+++ GoTvCore/GoTvP2P/NetServices/NetActionQueryMyIp.cpp
@@ -47,7 +47,7 @@ void NetActionQueryMyIp::doAction( void )
 	std::string strHttpUrl;
 	buildQueryMyIpUrl( netServConn, strHttpUrl );
 
-	RCODE rc = netServConn->sendData( strHttpUrl.c_str(), strHttpUrl.length() );
+	RCODE rc = netServConn->sendData( strHttpUrl.c_str(), (int)strHttpUrl.length() );
 	if( rc )
 	{
 		LogMsg( LOG_ERROR, "NetActionQueryMyIp::doAction: sendData error %d\n", rc );
